Rename again because lyric duplicated with exist lyric hit object.
