+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/LyricEditor.cs
@@ -8,7 +8,7 @@ using osu.Game.Rulesets.Objects;
 using osu.Game.Screens.Edit;
 using osu.Game.Skinning;
 
-namespace osu.Game.Rulesets.Karaoke.Edit.Lyric
+namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics
 {
     public class LyricEditor : Container
     {
